Write and read each byte separately in SPI driver
We wait for RXNE after each byte. This is slower since we are not
sending the next byte as soon as TXE is set. However, we make sure to
not miss any bytes by being to slow, which could make us stuck at a
busy loop. This ensures we are not failing if we receive an interrupt
somewhere during the transfer, and allows us to use a higher clock frequency.